+++ mse_cli/command/init.py
@@ -10,9 +10,9 @@ from mse_cli.log import LOGGER as LOG
 
 def add_subparser(subparsers):
     """Define the subcommand."""
-    parser = subparsers.add_parser("init",
-                                   help="create a new configuration file in "
-                                   "the current directory")
+    parser = subparsers.add_parser(
+        "init", help="create a new configuration file in " "the current directory"
+    )
 
     parser.set_defaults(func=run)
 
@@ -28,20 +28,22 @@ def run(_args):
     code_location = input("Code location [.]:") or "."
     python_application = input("Python application [app:app]: ") or "app:app"
     healthcheck_endpoint = input("Health check endpoint [/]: ") or "/"
-    secrets_enable = input(
-        "Do you have application secrets to provide [no]: ") or "no"
+    secrets_enable = input("Do you have application secrets to provide [no]: ") or "no"
     secrets = "secrets.json" if secrets_enable.lower() in ["y", "yes"] else None
 
-    app = AppConf(name=app_name,
-                  project=project_name,
-                  resource=resource,
-                  code=CodeConf(location=code_location,
-                                python_application=python_application,
-                                healthcheck_endpoint=healthcheck_endpoint,
-                                docker=docker,
-                                secrets=secrets))
+    app = AppConf(
+        name=app_name,
+        project=project_name,
+        resource=resource,
+        code=CodeConf(
+            location=code_location,
+            python_application=python_application,
+            healthcheck_endpoint=healthcheck_endpoint,
+            docker=docker,
+            secrets=secrets,
+        ),
+    )
 
     path = Path(os.getcwd())
     app.save(path)
-    LOG.success("Your app configuration has been saved in: %s",
-                path / "mse.toml")
+    LOG.success("Your app configuration has been saved in: %s", path / "mse.toml")